requirements.txt
pytest==9.0.2
selenium==4.41.0
requests==2.33.0
allure-pytest==2.15.3
webdriver-manager==4.0.2
python-dotenv==1.2.2
allure-pytest==2.15.3